+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Enhancement: Custom block example in React.js
 * Enhancement: JS example with REST and nonce
 * Enhancement: Page-Madness-Detector and Assets package added
+* Improved: Cli tests better example
 * Removed: Grunt replaced by Webpack
 
 ## 3.2.7